logTimeStamp/Jia: 1. 重構atom 2. 調整clearPageViewLog邏輯，計時器一起歸零
