+++ src/atoms/modalAtoms.ts
@@ -1,0 +1,6 @@
+import { atom } from 'jotai';
+
+const showSettingModalAtom = atom(false);
+const showAnalyticsModalAtom = atom(false);
+
+export { showSettingModalAtom, showAnalyticsModalAtom };
