+++ src/atoms/taskAtoms.ts
@@ -1,102 +1,13 @@
 import { atom } from 'jotai';
-import { z } from 'zod';
-
-// types
-type UserInfo = z.infer<typeof UserInfoSchema>;
-
-type Task = {
-  id: number;
-  taskName: string;
-  workTimeRef: number;
-  breakTimeRef: number;
-  pageViewLog?: PageViewLog[];
-};
-
-type PageViewLog = {
-  visible: boolean;
-  timestamp: number;
-};
-
-type ScheduledSegment = {
-  state: 'work' | 'break';
-  percent: number;
-  start: number;
-  end: number;
-};
-
-// schema
-const TaskSchema = z.object({
-  id: z.number(),
-  taskName: z.string(),
-  workTimeRef: z.number(),
-  breakTimeRef: z.number(),
-  pageViewLog: z
-    .array(
-      z.object({
-        visible: z.boolean(),
-        timestamp: z.number(),
-      })
-    )
-    .optional(),
-});
-
-const UserInfoSchema = z.object({
-  tasks: z.array(TaskSchema),
-  currentTaskId: z.number(),
-});
-
-// configure
-const STORAGE_KEY = 'userInfo';
-const DEFAULT_TASK = 'Time to focus!';
-const DEFAULT_WORKTIME = 25 * 60; // minutes
-const DEFAULT_BREAKTIME = 5 * 60; // minutes
-
-// default
-const defaultUserInfo: UserInfo = {
-  tasks: [],
-  currentTaskId: 0,
-};
-
-// state
-const timerStateAtom = atom<'work' | 'break'>('work');
-const isRunning = atom(false);
-const stopwatchAtom = atom(0);
-
-// function
-const getUserInfo = (): UserInfo => {
-  try {
-    const stored = localStorage.getItem(STORAGE_KEY);
-    if (!stored) return defaultUserInfo;
-    const parsedData = JSON.parse(stored);
-    return UserInfoSchema.parse(parsedData);
-  } catch (e) {
-    console.warn('Invalid user info in localStorage, resetting...', e);
-    localStorage.removeItem(STORAGE_KEY);
-    return defaultUserInfo;
-  }
-};
-
-const userInfoAtom = atom<UserInfo>(getUserInfo());
-const showSettingModalAtom = atom(false);
-const showAnalyticsModalAtom = atom(false);
-
-const currentTaskAtom = atom((get) => {
-  const userInfo = get(userInfoAtom);
-  const task = userInfo.tasks.find((task) => {
-    return task.id === userInfo.currentTaskId;
-  });
-  return {
-    taskName: task?.taskName || DEFAULT_TASK,
-    workTimeRef: task?.workTimeRef || DEFAULT_WORKTIME,
-    breakTimeRef: task?.breakTimeRef || DEFAULT_BREAKTIME,
-    pageViewLog: task?.pageViewLog || null,
-  };
-});
+import { Task, UserInfo, PageViewLog } from '@/types/taskTypes';
+import { STORAGE_KEY } from '@/constants/storage';
+import { userInfoAtom } from './userAtoms';
+import { showAnalyticsModalAtom } from './modalAtoms';
 
 const addTaskAtom = atom(
   null,
   (get, set, task: { taskName: string; workTimeRef: number; breakTimeRef: number }) => {
-    const prev = get(userInfoAtom);
+    const prev = get(userInfoAtom) as UserInfo;
     const newTask: Task = {
       id: Date.now(),
       taskName: task.taskName,
@@ -117,7 +28,7 @@ const addTaskAtom = atom(
 );
 
 const clearTaskAtom = atom(null, (get, set) => {
-  const userInfo = get(userInfoAtom);
+  const userInfo = get(userInfoAtom) as UserInfo;
   if (userInfo.currentTaskId === null) return;
   const updateData = {
     tasks: [],
@@ -129,7 +40,7 @@ const clearTaskAtom = atom(null, (get, set) => {
 });
 
 const clearPageViewLogAtom = atom(null, (get, set) => {
-  const userInfo = get(userInfoAtom);
+  const userInfo = get(userInfoAtom) as UserInfo;
   const updatedTasks = userInfo.tasks.map((task) => {
     if (task.id === userInfo.currentTaskId) {
       return {
@@ -150,7 +61,7 @@ const clearPageViewLogAtom = atom(null, (get, set) => {
 });
 
 const updatePageViewAtom = atom(null, (get, set, log: PageViewLog) => {
-  const userInfo = get(userInfoAtom);
+  const userInfo = get(userInfoAtom) as UserInfo;
   const updatedTasks = userInfo.tasks.map((task) => {
     if (task.id === userInfo.currentTaskId) {
       return {
@@ -169,19 +80,4 @@ const updatePageViewAtom = atom(null, (get, set, log: PageViewLog) => {
   set(userInfoAtom, updateData);
 });
 
-export {
-  timerStateAtom,
-  isRunning,
-  stopwatchAtom,
-  userInfoAtom,
-  currentTaskAtom,
-  addTaskAtom,
-  DEFAULT_WORKTIME,
-  DEFAULT_BREAKTIME,
-  showSettingModalAtom,
-  updatePageViewAtom,
-  showAnalyticsModalAtom,
-  clearPageViewLogAtom,
-  clearTaskAtom,
-};
-export type { PageViewLog, ScheduledSegment };
+export { addTaskAtom, clearTaskAtom, updatePageViewAtom, clearPageViewLogAtom };
