+++ src/constants/storage.ts
@@ -1,0 +1,6 @@
+const STORAGE_KEY = 'userInfo';
+const DEFAULT_TASK = 'Time to focus!';
+const DEFAULT_WORKTIME = 25 * 60; // minutes
+const DEFAULT_BREAKTIME = 5 * 60; // minutes
+
+export { STORAGE_KEY, DEFAULT_TASK, DEFAULT_WORKTIME, DEFAULT_BREAKTIME };
